test: extract shared FakeMarketPriceServiceFacade and offer fixtures
